refactor(registry): split resolve versus refresh
